requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy 
mysql-connector
globals